+++ Cargo.toml
@@ -27,6 +27,7 @@ members = [
     "crates/perl-lsp-transport",
     "crates/perl-lsp-tooling",
     "crates/perl-subprocess-runtime",
+    "crates/perl-lsp-formatting-types",
     "crates/perl-lsp-formatting",
     "crates/perl-lsp-diagnostics",
     "crates/perl-lsp-semantic-tokens",
@@ -145,6 +146,7 @@ allow = [
     # Tier 4 — LSP providers
     "perl-lsp-navigation",
     "perl-lsp-tooling",
+    "perl-lsp-formatting-types",
     "perl-lsp-formatting",
     "perl-lsp-semantic-tokens",
     "perl-lsp-providers",
@@ -264,6 +266,7 @@ perl-lsp-capability-map = { path = "crates/perl-lsp-capability-map", version = "
 perl-parser-core = { path = "crates/perl-parser-core", version = "0.10.0" }
 perl-lsp-transport = { path = "crates/perl-lsp-transport", version = "0.10.0" }
 perl-lsp-tooling = { path = "crates/perl-lsp-tooling", version = "0.10.0" }
+perl-lsp-formatting-types = { path = "crates/perl-lsp-formatting-types", version = "0.10.0" }
 perl-lsp-formatting = { path = "crates/perl-lsp-formatting", version = "0.10.0" }
 perl-tdd-support = { path = "crates/perl-tdd-support", version = "0.10.0" }
 perl-lsp-diagnostics = { path = "crates/perl-lsp-diagnostics", version = "0.10.0" }
