+++ crates/perl-lsp-formatting-types/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "perl-lsp-formatting-types"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+description = "Shared formatting DTOs for perl-lsp formatting workflows"
+readme = "README.md"
+license = "MIT OR Apache-2.0"
+repository.workspace = true
+homepage.workspace = true
+documentation = "https://docs.rs/perl-lsp-formatting-types"
+keywords = ["perl", "lsp", "formatting", "text-edit"]
+categories = ["development-tools", "text-editors"]
+
+[lib]
+doctest = false
+
+[dependencies]
+serde = { version = "1.0.228", features = ["derive"] }
+
+[lints]
+workspace = true
