+++ crates/perl-lsp-formatting/Cargo.toml
@@ -18,7 +18,7 @@ doctest = false
 
 [dependencies]
 perl-lsp-tooling = { workspace = true }
-serde = { version = "1.0.228", features = ["derive"] }
+perl-lsp-formatting-types = { workspace = true }
 thiserror = "2.0"
 
 [dev-dependencies]
